chapter11 Linux 컴파일 오류 수정: stdint.h 추가, unsigned __int64 → unsigned long long

+++ chapter11/05arithptr.c
@@ -8,8 +8,8 @@ int main(void)
 	//pd = 100;					// double 포인터에 100으로 저장하면 경고 발생
 
 	printf("%lld  %lld  %lld\n", (long long)(pc - 1), (long long)pc, (long long)(pc + 1));
-	printf("%llu  %llu  %llu\n", (unsigned __int64)(pi - 1),
-		                         (unsigned __int64)pi, (unsigned __int64)(pi + 1));
+	printf("%llu  %llu  %llu\n", (unsigned long long)(pi - 1),
+		                         (unsigned long long)pi, (unsigned long long)(pi + 1));
 	printf("%llu  %llu  %llu\n", (unsigned long long)(pd - 1),
 		                         (unsigned long long)pd, (unsigned long long)(pd + 1));
